project/requirements.txt
fastapi
uvicorn
requests
python-multipart
websockets
python-dotenv
pydantic[email]

PyJWT
pydantic-settings

sqlalchemy
aiosqlite
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
docx2txt

alembic
aiologger
aiofiles

passlib[bcrypt]